rm FloatTensor

+++ data/collate.py
@@ -1,4 +1,5 @@
 import torch
+import numpy as np
 
 def detection_collate(batch):
     imgs = []
@@ -7,4 +8,4 @@ def detection_collate(batch):
         imgs.append(torch.from_numpy(img))
         targets.append(torch.from_numpy(target))
 
-    return torch.stack(imgs, 0), torch.FloatTensor(targets)
+    return torch.stack(imgs, 0), targets
